apps/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
alpaca-py>=0.21.0
pytz>=2024.1
websockets>=12.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
pandas>=2.0.0
